# apps/backend/openapi/openapi.yaml
openapi: 3.0.3 # This matches the pattern
info:
  title: Circa
  description: API Specification for Circa (Ajọra) - private rotating savings groups with wallet-based auth.
  version: 0.0.1
...
servers:
  - url: http://localhost:8081
    description: Main API server
security:
  - SessionAuth: []
paths:
...
  /auth/signup:
    post:
      tags: [auth]
      summary: Sign up a new user
      operationId: authSignup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthSignupRequest"
      responses:
        "200":
          description: Signup successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSignupResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBadRequest"
  /auth/login:
    post:
      tags: [auth]
      summary: Request login magic link
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
      responses:
        "200":
          description: Login link sent (if account exists)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthLoginResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBadRequest"
  /auth/verify:
    post:
      tags: [auth]
      summary: Verify email token create a session (sets HttpOnly cookie)
      operationId: authVerify
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthVerifyRequest"
      responses:
        "200":
          description: Signed in successfully (session cookie set)
          headers:
            Set-Cookie:
              description: HttpOnly session cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthVerifyResponse"
        "400":
          description: Bad Request (invalid payload, expired/used nonce, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBadRequest"
        "401":
          description: Unauthorized (signature invalid)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorUnauthorized"
  /auth/signup/complete:
    post:
      tags: [auth]
      summary: Complete signup with wallet signature (creates user and main session)
      operationId: authSignupComplete
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthVerifyWalletRequest"
      responses:
        "200":
          description: Signup completed successfully (main session cookie set)
          headers:
            Set-Cookie:
              description: HttpOnly session cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthVerifyWalletResponse"
        "400":
          description: Bad Request (invalid payload, expired/used nonce, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBadRequest"
        "401":
          description: Unauthorized (invalid session, signature, or nonce)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorUnauthorized"
        "409":
          description: Conflict (wallet or email already in use)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBadRequest"
  /auth/nonce:
    post:
      tags: [auth]
      summary: Request a nonce for wallet sign-in
      operationId: authNonce
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthNonceRequest"
      responses:
        "200":
          description: Nonce issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthNonceResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBadRequest"
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: circa_session
  schemas:
    UUID:
      type: string
      format: uuid
    Address:
      type: string
      description: EVM address (0x-prefixed, 40 hex chars)
      pattern: "^0x[a-fA-F0-9]{40}$"
    ChainId:
      type: integer
      minimum: 1
      description: EVM chain id
    Timestamp:
      type: string
      format: date-time
    ErrorBadRequest:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: Bad request
    ErrorUnauthorized:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 401
        message:
          type: string
          example: Unauthorized
...
    AuthSignupRequest:
      type: object
      required: [full_name, email, display_name]
      properties:
        full_name:
          type: string
          minLength: 1
          maxLength: 50
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 255
        display_name:
          type: string
          minLength: 1
          maxLength: 50
          nullable: true
      additionalProperties: false
    AuthSignupResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
      additionalProperties: false
    AuthLoginRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 255
      additionalProperties: false
    AuthLoginResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
      additionalProperties: false
    AuthNonceRequest:
      type: object
      required: [address]
      properties:
        address:
          $ref: "#/components/schemas/Address"
        chainId:
          $ref: "#/components/schemas/ChainId"
      additionalProperties: false
    AuthNonceResponse:
      type: object
      required: [nonce, expiresAt]
      properties:
        nonce:
          type: string
          description: One-time nonce used for wallet sign-in (hex-encoded)
          pattern: "^0x[a-fA-F0-9]+$"
        expiresAt:
          $ref: "#/components/schemas/Timestamp"
        messageTemplate:
          type: string
          description: Optional template the client can use to construct the sign-in message
    AuthVerifyRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: Token generated during signup/signin
      additionalProperties: false
    AuthVerifyResponse:
      type: object
      required: [email, needsWallet]
      properties:
        email:
          type: string
          format: email
        displayName:
          type: string
          nullable: true
        needsWallet:
          type: boolean
          description: Whether the user needs to connect a wallet to complete signup
    AuthVerifyWalletRequest:
      type: object
      required: [address, signature, message]
      properties:
        address:
          $ref: "#/components/schemas/Address"
        signature:
          type: string
          description: ECDSA signature of the message (hex-encoded, 0x-prefixed)
          pattern: "^0x[a-fA-F0-9]{130}$"
        message:
          type: string
          description: The exact message that was signed (must include nonce and domain binding)
      additionalProperties: false
    AuthVerifyWalletResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    # -----------------------------
    # USER / PROFILE
    # -----------------------------
    User:
      type: object
      required: [id, address, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        address:
          $ref: "#/components/schemas/Address"
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          format: uri
          nullable: true
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
          nullable: true
...